feat: add step limit to prevent infinite loops
On Linux, when uxnemu enters an infinite loop, the only way to stop
it is to send SIGKILL to it. This makes debugging infinite loops very
annoying. This patch adds a step limit per execution, which should be
well out of the realm of a plausible use, and helps stop infinite
loops.